billing-stocks-management-system-be/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
pillow==10.1.0
pytesseract==0.3.10
google-cloud-vision==3.4.5
google-cloud-translate==3.12.1
SpeechRecognition==3.10.0
pandas==2.1.1
openpyxl==3.1.2